% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jvm_lifecycle.R
\name{checkJarLibAvailable}
\alias{checkJarLibAvailable}
\title{Check whether the eGPS jar libraries are available.}
\usage{
checkJarLibAvailable(programPath = NA_character_)
}
\arguments{
\item{programPath}{Optional runtime path.}
}
\value{
TRUE if all required files exist, FALSE otherwise.
}
\description{
Check whether the eGPS jar libraries are available.
}
