% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI_invoker_forR.R
\name{correlationVis_expressionProfile}
\alias{correlationVis_expressionProfile}
\title{Correlation Visualization with the expression profile}
\usage{
correlationVis_expressionProfile(
  egps,
  df,
  expMatrixCol = c("name"),
  cat1Col = c("name", "cat1"),
  cat2Col = c("name", "cat2", "value")
)
}
\arguments{
\item{egps}{the eGPS Java instance}

\item{df}{the data frame}

\item{expMatrixCol}{the col names of the expression matrix, only the first 'name' column is needed.}

\item{cat1Col}{the col names of the category one: name, cat1, or color(optional)}

\item{cat2Col}{the col names of the category two: name, cat2 and value}
}
\value{
no return
}
\description{
Correlation Visualization with the expression profile
}
\examples{
\dontrun{

egps <- launchEGPS_withinR(autoSetWD = T)

compsLength <- 10
sampleSize <- 8
gNames <- paste0('Gene', 1:compsLength)
simulatedData <- setNames(
  as.data.frame(replicate(sampleSize, runif(compsLength, 0, sampleSize))),
  paste0('S', 1:sampleSize)
)
df <- data.frame(name = gNames, simulatedData,
                      cat1 = rep(c("Cat1","Cat2","Cat3"), each = 3, length.out = compsLength),
                      cat2 = rep(c("Cat1","Cat2","Cat3"), each = 4, length.out = compsLength),
                      value = 1)
# make sure the column names are matched
correlationVis_expressionProfile(egps, df,expMatrixCol = c("name"),
                                 cat1Col = c("name", "cat1"),
                                 cat2Col = c("name", "cat2", "value"))

}
}
