% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.coerceStoredVars}
\alias{.coerceStoredVars}
\title{Coerce stored vars to a list.}
\usage{
.coerceStoredVars(vars)
}
\arguments{
\item{vars}{the stored vars (list or environment)}
}
\value{
the vars coerced to a list, or NULL
}
\description{
Coerce stored vars to a list.
}
\keyword{internal}
