% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.normalizePathForJava}
\alias{.normalizePathForJava}
\title{Normalize a file path for Java compatibility.}
\usage{
.normalizePathForJava(path, mustWork = TRUE)
}
\arguments{
\item{path}{the file path to normalize}

\item{mustWork}{logical, should the path exist?}
}
\value{
the normalized path with forward slashes
}
\description{
Normalize a file path for Java compatibility.
}
\keyword{internal}
