% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamental_evoltre.R
\name{evoltre_getNodeNames}
\alias{evoltre_getNodeNames}
\title{With importing the tree_path and the onlyLeaf parameters, return the names array.}
\usage{
evoltre_getNodeNames(tree_path, onlyLeaf = FALSE)
}
\arguments{
\item{tree_path}{The phylogenetic tree path}

\item{onlyLeaf}{only leaf names or the total names (including internal nodes)}
}
\value{
the string vecotr of names.
}
\description{
With importing the tree_path and the onlyLeaf parameters, return the names array.
}
\examples{
\dontrun{
evoltre_getNodeNames('path/to/tree.nwk' ,  onlyLeaf = F);
}

}
