% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_fastadumper.R
\name{fastadumper_partialMatch}
\alias{fastadumper_partialMatch}
\title{Obtain sequences from the fasta file}
\usage{
fastadumper_partialMatch(
  fastaPath,
  entries,
  outPath = "extracted.seq.fa",
  keepSearch = F
)
}
\arguments{
\item{fastaPath}{the input path of the fasta}

\item{entries}{the string vector of the sequence to extract}

\item{outPath}{the output path}

\item{keepSearch}{whether repeat search for the entries}
}
\value{
nothing
}
\description{
Run example
# entry = c("gene1","gene2","gene2"))
# fastadumper_partialMatch(fastaPath = "subject.fa",entries = entry, outPath = "out.fa")
}
\details{
Note: the fasta file must be ContinuousFasta.
}
