% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_homoidentify.R
\name{hmmer_domtbl2tsv}
\alias{hmmer_domtbl2tsv}
\title{Parse the hmmer domtbl output to tsv file}
\usage{
hmmer_domtbl2tsv(hmmerDomtbl, outputPath)
}
\arguments{
\item{hmmerDomtbl}{the path of hmmer domtbl}

\item{outputPath}{the path of output file}
}
\value{
nothing
}
\description{
Parse the hmmer domtbl output to tsv file
}
