% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initializeJVM4eGPS}
\alias{initializeJVM4eGPS}
\title{Initialize the JVM, if the JVM is already launched, do nothing.}
\usage{
initializeJVM4eGPS()
}
\value{
nothing, error will throw if the command not meet.
}
\description{
Initialize the JVM, if the JVM is already launched, do nothing.
}
