% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_launch.R
\name{launchEGPSDesktop}
\alias{launchEGPSDesktop}
\title{Launch the eGPS desktop and return the bridge instance.}
\usage{
launchEGPSDesktop()
}
\value{
The Java bridge object used by R4eGPS.
}
\description{
Launch the eGPS desktop and return the bridge instance.
}
