% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list2jsonStr}
\alias{list2jsonStr}
\title{Convert the R object, usually the list inside data.frames, to JSON string with some customized configurations/rules.}
\usage{
list2jsonStr(ll)
}
\arguments{
\item{ll}{a list, usually with data.frame inside it}
}
\value{
the JSON string
}
\description{
Convert the R object, usually the list inside data.frames, to JSON string with some customized configurations/rules.
}
