% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_launch.R
\name{openModernTreeViewFromConfig}
\alias{openModernTreeViewFromConfig}
\title{Open Modern Tree View from a config file.}
\usage{
openModernTreeViewFromConfig(configPath)
}
\arguments{
\item{configPath}{A VOICE config file path.}
}
\value{
Invisibly returns the Java bridge instance.
}
\description{
Open Modern Tree View from a config file.
}
