% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_launch.R
\name{openPathwayFamilyBrowserFromConfig}
\alias{openPathwayFamilyBrowserFromConfig}
\title{Open Pathway Family Browser from a config file.}
\usage{
openPathwayFamilyBrowserFromConfig(configPath)
}
\arguments{
\item{configPath}{A VOICE config file path.}
}
\value{
Invisibly returns the Java bridge instance.
}
\description{
Open Pathway Family Browser from a config file.
}
