% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jvm_lifecycle.R
\name{setLib_and_launchJVM}
\alias{setLib_and_launchJVM}
\title{Set library and launch JVM against the configured runtime.}
\usage{
setLib_and_launchJVM()
}
\value{
The effective Java classpath entries.
}
\description{
Set library and launch JVM against the configured runtime.
}
