% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_structdraw.R
\name{structDraw_multi_genes}
\alias{structDraw_multi_genes}
\title{Draw the multi genes struct}
\usage{
structDraw_multi_genes(list, width = 800, height = 500, gHeight = 50)
}
\arguments{
\item{list}{list of the ploted views}

\item{width}{the width of the displayed form view}

\item{height}{the height of the displayed form view}

\item{gHeight}{the height of each gene}
}
\value{
no return
}
\description{
Draw the multi genes struct
}
\examples{
\dontrun{
list1 <- list(
gene1 = list(length = 250, start = c(1,10,101,200), end = c(8, 56, 152, 230), color = c("#E63946", "#457B9D", "#2A9D8F", "#F4A261")),
 gene2 = list(length = 350,start = c(1,180,210,261), end = c(150, 200, 250, 300), color = c("#E63946", "#457B9D", "#2A9D8F", "#F4A261")),
 gene3 = list(length = 300,start = c(51,101,200), end = c(100,150,231), color = c("#2A9D8F", "#F4A261", "#1D3557"))
)
list2 <- list(
 gene1 = list(length = 100, start = 1, end = 50, color = c("#E63946")),
 gene3 = list(length = 300,start = c(51,101), end = c(100,150), color = c("#2A9D8F", "#F4A261"))
)
R4eGPS::structDraw_multi_genes(list1)
}
}
