% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_taxonomy.R
\name{taxonomy_getRankedLineages}
\alias{taxonomy_getRankedLineages}
\title{Get the ranked lineages of a species}
\usage{
taxonomy_getRankedLineages(filePath, entries)
}
\arguments{
\item{filePath}{the NCBI ranked lineages dump file}

\item{entries}{the taxon ids}
}
\value{
the string double array
}
\description{
Get the ranked lineages of a species
}
\examples{
taxonomy_getRankedLineages("path", c(1,2))
}
