% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writeDataFrameToTsv}
\alias{writeDataFrameToTsv}
\title{Write a data.frame to a tab-separated file.}
\usage{
writeDataFrameToTsv(df, path)
}
\arguments{
\item{df}{the data.frame}

\item{path}{the output path}
}
\value{
nothing
}
\description{
Write a data.frame to a tab-separated file.
}
