% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamental_evoltre.R
\name{evoltre_getNodeNames}
\alias{evoltre_getNodeNames}
\title{With importing the tree_path and the onlyLeaf parameters, return the names array.}
\usage{
evoltre_getNodeNames(
  tree_path,
  targetHTU = NULL,
  getOTU = TRUE,
  getHTU = FALSE
)
}
\arguments{
\item{tree_path}{The phylogenetic tree path}

\item{targetHTU}{NULL for the root node, else input the internal node name}

\item{getOTU}{whether get the OTU names}

\item{getHTU}{whether get the HTU names}
}
\value{
the string vecotr of names.
}
\description{
With importing the tree_path and the onlyLeaf parameters, return the names array.
}
\examples{
\dontrun{
## get the leaf names:
evoltre_getNodeNames('path/to/tree.nwk' ,  getOTU = T, getHTU = F);
## get the leaf names of certain internal node
evoltre_getNodeNames('path/to/tree.nwk' , targetHTU = 'name1', getOTU = T, getHTU = F);
}

}
