% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Java_interaction.R
\name{launchEGPS_withinR}
\alias{launchEGPS_withinR}
\title{Launch eGPS2 GUI desktop in R}
\usage{
launchEGPS_withinR(programPath = NA_character_)
}
\arguments{
\item{programPath}{the eGPS2 software program path, the R4eGPS will remember the config parameter. This is setting for the first time, for the second time or later, this is not need to do. You can also set via 'setGlobalVars(list(eGPS_software_path = "/path/to/software"))'}
}
\value{
the JAVA object of eGPS2
}
\description{
If you want to see the JVM parameters, using options(java.parameters = c("-Xmx4G"))
}
\examples{
\dontrun{
# This is for the first time
egps <- launchEGPS_withinR(programPath = '/path/to/egps/')
# routine use
egps <- launchEGPS_withinR()
egps$callTest("Hello eGPS in R!")
}
}
