% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{setGlobalVars}
\alias{setGlobalVars}
\title{set The global Vars.
Current support is:}
\usage{
setGlobalVars(varList)
}
\arguments{
\item{varList}{a named vector with string key-values.}
}
\value{
no return
}
\description{
'eGPS_software_path': the eGPS jars lib path, for example: "C:/Users/yudal/Documents/project/eGPS2/eGPS_v2_windows_64bit/eGPS_lib"
}
\examples{
\dontrun{

setGlobalVars( list( eGPS_software_path = '/your/path/dir') )
vars <- getGlobalVars();vars[['eGPS_software_path']] <- "/new/path/dir"

}
}
