% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Java_interaction.R
\name{setLib_and_launchJVM}
\alias{setLib_and_launchJVM}
\title{Set library and launch JVM}
\usage{
setLib_and_launchJVM()
}
\value{
the lib paths
}
\description{
Set library and launch JVM
}
\examples{
setLib_and_launchJVM()
}
