% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writeDataFrameToTsv}
\alias{writeDataFrameToTsv}
\title{Function as its name.
how to use: writeDataFrameToTsv(data.frame(), "/output/path")}
\usage{
writeDataFrameToTsv(df, path)
}
\arguments{
\item{df}{the data.frame}

\item{path}{the output path}
}
\value{
nothing
}
\description{
importFrom("utils", "write.table")
}
